+++ src/misc/math.hpp
@@ -1,5 +1,18 @@
 #pragma once
 
+// NOLINTBEGIN(readability-identifier-naming)
+
+#define Mathf_SQRT12 ((float)Math_SQRT12)
+#define Mathf_SQRT2 ((float)Math_SQRT2)
+#define Mathf_LN2 ((float)Math_LN2)
+#define Mathf_PI ((float)Math_PI)
+#define Mathf_TAU ((float)Math_TAU)
+#define Mathf_E ((float)Math_E)
+#define Mathf_INF ((float)Math_INF)
+#define Mathf_NAN ((float)Math_NAN)
+
+// NOLINTEND(readability-identifier-naming)
+
 namespace godot::Math {
 
 _FORCE_INLINE_ void decompose(Basis& p_basis, Vector3& p_scale) {